requirements_standalone.txt
Flask>=2.3.0
opencv-python>=4.8.0
numpy>=1.24.0
astropy>=5.3.0
Pillow>=10.0.0
pywebview>=4.4.0

